w3-Assignment/requirements.txt
google-generativeai
streamlit
python-dotenv
langchain_core
config